Downgrading several fatal messages to allow extraction
My file has 0x10 headers on all samples. That is not fatal

Ten of 4000 samples has an unknown crc 0x8d00698d. That only applies to those samples that may or may not be playable with an arbitrary header applied. The crc may be in https://github.com/HearthSim/python-fsb5/blob/master/fsb5/vorbis_headers.py

The test file is speech-english.fsb and can be obtained from steamcmd download_depot 378540 378544